converting registers to a dictionary
day 9 requires an infinitely long tape, which is better handled by a map
benefit is we don't have to extend anymore!
downside is, we might have to initialize values that aren't created yet
all this is done in parse_parameter!

day 9 part 1 examples pass
output is incorrect due to a store to the relative base
tests pass for previous days as well